% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statascii-package.R
\docType{package}
\name{statascii-package}
\alias{statascii}
\alias{statascii-package}
\title{statascii: Create Stata-like Tables in the R Console}
\description{
Create Stata-like tables in the R console, including right-aligned text, Stata-like padding, various Stata table formats (`oneway`, `twoway`, and `summary`), and Unicode box drawing characters. 'statascii()' borrows heavily from 'asciify()'. The 'asciify()' function was written by @gavinsimpson in StackOverflow (https://stackoverflow.com/questions/13011383) and GitHub Gist (https://gist.github.com/gavinsimpson). The 'statascii()' function was first written by @gvelasq2 in Github Gist (https://gist.github.com/gvelasq2).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gvelasq2/statascii}
  \item Report bugs at \url{https://github.com/gvelasq2/statascii/issues}
}

}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github at gmail.com}

}
\keyword{internal}
